// src/lib/types/ast.d.ts
interface HasAttributes {
    attributes?: Attributes;
    pos?: Pos;
}
interface HasChildren<A> {
    children: A[];
}
interface HasText {
    text: string;
}
type Block = Para | Heading | ThematicBreak | Section | Div | CodeBlock | RawBlock | BlockQuote | OrderedList | BulletList | TaskList | DefinitionList | Table;
interface Para extends HasAttributes {
    tag: "para";
    children: Inline[];
}
interface Heading extends HasAttributes {
    tag: "heading";
    level: number;
    children: Inline[];
}
interface ThematicBreak extends HasAttributes {
    tag: "thematic_break";
}
interface Section extends HasAttributes {
    tag: "section";
    children: Block[];
}
interface Div extends HasAttributes {
    tag: "div";
    children: Block[];
}
interface BlockQuote extends HasAttributes {
    tag: "block_quote";
    children: Block[];
}
interface CodeBlock extends HasAttributes {
    tag: "code_block";
    lang?: string;
    text: string;
}
interface RawBlock extends HasAttributes {
    tag: "raw_block";
    format: string;
    text: string;
}
interface BulletList extends HasAttributes {
    tag: "bullet_list";
    tight: boolean;
    style: BulletListStyle;
    children: ListItem[];
}
type BulletListStyle = "+" | "-" | "*";
interface ListItem extends HasAttributes {
    tag: "list_item";
    children: Block[];
}
interface TaskList extends HasAttributes {
    tag: "task_list";
    tight: boolean;
    children: TaskListItem[];
}
interface TaskListItem extends HasAttributes {
    tag: "task_list_item";
    checkbox: CheckboxStatus;
    children: Block[];
}
type CheckboxStatus = "checked" | "unchecked";
interface OrderedList extends HasAttributes {
    tag: "ordered_list";
    style: OrderedListStyle;
    tight: boolean;
    start?: number;
    children: ListItem[];
}
type OrderedListStyle = "1." | "1)" | "(1)" | "a." | "a)" | "(a)" | "A." | "A)" | "(A)" | "i." | "i)" | "(i)" | "I." | "I)" | "(I)";
interface Caption extends HasAttributes {
    tag: "caption";
    children: Inline[];
}
interface Table extends HasAttributes {
    tag: "table";
    children: [Caption, ...Row[]];
}
type Inline = Str | SoftBreak | HardBreak | NonBreakingSpace | Symb | Verbatim | RawInline | InlineMath | DisplayMath | Url | Email | FootnoteReference | SmartPunctuation | Emph | Strong | Link | Image | Span | Mark | Superscript | Subscript | Insert | Delete | DoubleQuoted | SingleQuoted;
interface Str extends HasAttributes {
    tag: "str";
    text: string;
}
interface FootnoteReference extends HasAttributes {
    tag: "footnote_reference";
    text: string;
}
type SmartPunctuationType = "left_single_quote" | "right_single_quote" | "left_double_quote" | "right_double_quote" | "ellipses" | "em_dash" | "en_dash";
interface SmartPunctuation extends HasAttributes {
    tag: "smart_punctuation";
    type: SmartPunctuationType;
    text: string;
}
interface SoftBreak extends HasAttributes {
    tag: "soft_break";
}
interface HardBreak extends HasAttributes {
    tag: "hard_break";
}
interface NonBreakingSpace extends HasAttributes {
    tag: "non_breaking_space";
}
interface Symb extends HasAttributes {
    tag: "symb";
    alias: string;
}
interface Verbatim extends HasAttributes {
    tag: "verbatim";
    text: string;
}
interface RawInline extends HasAttributes {
    tag: "raw_inline";
    format: string;
    text: string;
}
interface InlineMath extends HasAttributes {
    tag: "inline_math";
    text: string;
}
interface DisplayMath extends HasAttributes {
    tag: "display_math";
    text: string;
}
interface Url extends HasAttributes {
    tag: "url";
    text: string;
}
interface Email extends HasAttributes {
    tag: "email";
    text: string;
}
interface Link extends HasAttributes {
    tag: "link";
    destination?: string;
    reference?: string;
    children: Inline[];
}
interface Image extends HasAttributes {
    tag: "image";
    destination?: string;
    reference?: string;
    children: Inline[];
}
interface Emph extends HasAttributes {
    tag: "emph";
    children: Inline[];
}
interface Strong extends HasAttributes {
    tag: "strong";
    children: Inline[];
}
interface Span extends HasAttributes {
    tag: "span";
    children: Inline[];
}
interface Mark extends HasAttributes {
    tag: "mark";
    children: Inline[];
}
interface Superscript extends HasAttributes {
    tag: "superscript";
    children: Inline[];
}
interface Subscript extends HasAttributes {
    tag: "subscript";
    children: Inline[];
}
interface Delete extends HasAttributes {
    tag: "delete";
    children: Inline[];
}
interface Insert extends HasAttributes {
    tag: "insert";
    children: Inline[];
}
interface DoubleQuoted extends HasAttributes {
    tag: "double_quoted";
    children: Inline[];
}
interface SingleQuoted extends HasAttributes {
    tag: "single_quoted";
    children: Inline[];
}
interface DefinitionList extends HasAttributes {
    tag: "definition_list";
    children: DefinitionListItem[];
}
interface DefinitionListItem extends HasAttributes {
    tag: "definition_list_item";
    children: [Term, Definition];
}
interface Term extends HasAttributes {
    tag: "term";
    children: Inline[];
}
interface Definition extends HasAttributes {
    tag: "definition";
    children: Block[];
}
interface Row extends HasAttributes {
    tag: "row";
    head: boolean;
    children: Cell[];
}